requirements.txt
accelerate==1.0.1
datasets==3.0.0
fastchat==0.1.0
intel_extension_for_pytorch==2.5.0
matplotlib==3.10.0
modelscope==1.21.1
nanogcg==0.2.2
nltk==3.9.1
numpy==1.26.3
openai==1.45.1
pandas==2.2.2
peft==0.14.0
Pillow
psutil==6.0.0
python-dotenv==1.0.1
Requests==2.32.3
setuptools==72.1.0
tenacity==9.0.0
together==1.3.5
# torch_npu==2.4.0
tqdm==4.66.5
transformers==4.45.1
trl==0.11.3
vertexai==1.71.1
httpx==0.27.2 # needed lower version to work with openrouter (7 Jan 2025)
math-verify[antlr4_13_2]==0.7.0
# torch==2.4.0+cu124
